just do one thing make a drop-down where 3 option are there  setting , logout and billing on top menu on left side.
and in billing section show him bill pending and bifergation of the amount and allow user to download the invoice and make sure that invoice is looking professional. and to show the breakdown of the bill use accordian

for now make a payment btn as user clicks on it make a alert("clicked and its id").

and to make this  use shadcn only.